synvert-snippets Removing debug_me calls
debug_me  A tool to print the labeled value of variables.
          |__ http://github.com/MadBomber/debug_me

Lets resolve not to keep debugging statement in a
production code base.  This is my gem.  I use it all
the time ... and yes, occassionally I have left debug_me
calls in my commits.  BUT NOW, with synvert I can
configure a pre-commit hook to remove those calls.